requirements.txt
streamlit>=1.28.0
groq>=0.4.0
PyPDF2>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0